update files except main.cpp

+++ cutoff.h
@@ -3,4 +3,4 @@
 #include "atom.h"
 #include "Graph.cpp"
 
-Graph Cutoff(Molecule molecule, float rc);
+Graph Cutoff(Molecule molecule, double rc);
